+++ examples/simulation-study/src/llhd-profiles.R
@@ -1,6 +1,7 @@
 library(ggplot2)
 library(dplyr)
 library(stringr)
+library(gridExtra)
 
 INPUT_FILE <- "out/config.json"
 
@@ -32,8 +33,7 @@ x$neg_binom_p <- 1 - nb_params[,2] # R and Wikipedia have different parameterisa
 font_scale_factor <- 0.5
 fig_theme <- theme(
     axis.title.x = element_text(size = font_scale_factor * 22),
-    axis.title.y = element_blank(),
-    ## axis.text.y = element_blank(),
+    axis.title.y = element_text(size = font_scale_factor * 22),
     axis.ticks = ggplot2::element_blank(),
     axis.line.x.bottom = element_line(colour = "#000000"),
     legend.title = element_text(size = font_scale_factor * 22),
@@ -44,88 +44,99 @@ fig_theme <- theme(
     panel.grid.major.x = ggplot2::element_blank(),
     panel.background = ggplot2::element_blank(),
     strip.background = ggplot2::element_rect(fill="white"))
+
+y_scaling <- scale_y_continuous(breaks = seq(from = -60, to = -45, by = 5))
+
 truth_linetype <- "dashed"
 
 lambda_figure <- ggplot(filter(x, mu == PARAMS$mu, psi == PARAMS$psi, rho == PARAMS$rho, omega == PARAMS$omega, nu == PARAMS$nu), aes(x = lambda, y = llhd)) +
     geom_line() +
     geom_vline(xintercept = PARAMS$lambda, linetype = truth_linetype) +
     labs(x = "Birth rate",
-         y = "Log-Likelihood",
-         title = "Log-likelihood profile",
-         subtitle = "Known observation model parameters") +
+         y = "Log-Likelihood") +
+    y_scaling +
     fig_theme
 
 if (save_figures) {
-    ggsave("out/llhd-profile-lambda.pdf", plot = lambda_figure, height = 10.5, width = 14.8, units = "cm")
+    ggsave("out/llhd-profile-lambda.pdf",
+           plot = lambda_figure + labs(title = "Log-likelihood profile", subtitle = "Known observation model parameters"),
+           height = 10.5, width = 14.8, units = "cm")
 }
 
 mu_figure <- ggplot(filter(x, lambda == PARAMS$lambda, psi == PARAMS$psi, rho == PARAMS$rho, omega == PARAMS$omega, nu == PARAMS$nu), aes(x = mu, y = llhd)) +
     geom_line() +
     geom_vline(xintercept = PARAMS$mu, linetype = truth_linetype) +
     labs(x = "Death rate",
-         y = "Log-Likelihood",
-         title = "Log-likelihood profile",
-         subtitle = "Known observation model parameters") +
+         y = "Log-Likelihood") +
+    y_scaling +
     fig_theme
 
 if (save_figures) {
-    ggsave("out/llhd-profile-mu.pdf", plot = mu_figure, height = 10.5, width = 14.8, units = "cm")
+    ggsave("out/llhd-profile-mu.pdf",
+           plot = mu_figure + labs(title = "Log-likelihood profile", subtitle = "Known observation model parameters"),
+           height = 10.5, width = 14.8, units = "cm")
 }
 
 psi_figure <- ggplot(filter(x, lambda == PARAMS$lambda, mu == PARAMS$mu, rho == PARAMS$rho, omega == PARAMS$omega, nu == PARAMS$nu), aes(x = psi, y = llhd)) +
     geom_line() +
     geom_vline(xintercept = PARAMS$psi, linetype = truth_linetype) +
     labs(x = "Sampling rate",
-         y = "Log-Likelihood",
-         title = "Log-likelihood profile",
-         subtitle = "Known observation model parameters") +
+         y = "Log-Likelihood") +
+    y_scaling +
     fig_theme
 
 if (save_figures) {
-    ggsave("out/llhd-profile-psi.pdf", plot = psi_figure, height = 10.5, width = 14.8, units = "cm")
+    ggsave("out/llhd-profile-psi.pdf",
+           plot = psi_figure + labs(title = "Log-likelihood profile", subtitle = "Known observation model parameters"),
+           height = 10.5, width = 14.8, units = "cm")
 }
 
 rho_figure <- ggplot(filter(x, lambda == PARAMS$lambda, mu == PARAMS$mu, psi == PARAMS$psi, omega == PARAMS$omega, nu == PARAMS$nu), aes(x = rho, y = llhd)) +
     geom_line() +
     geom_vline(xintercept = PARAMS$rho, linetype = truth_linetype) +
-    labs(x = "Catastrophe extinction probability",
-         y = "Log-Likelihood",
-         title = "Log-likelihood profile",
-         subtitle = "Known observation model parameters") +
+    labs(x = "Catastrophe extinction\nprobability",
+         y = "Log-Likelihood") +
+    y_scaling +
     fig_theme
 
 if (save_figures) {
-    ggsave("out/llhd-profile-rho.pdf", plot = rho_figure, height = 10.5, width = 14.8, units = "cm")
+    ggsave("out/llhd-profile-rho.pdf",
+           plot = rho_figure + labs(title = "Log-likelihood profile", subtitle = "Known observation model parameters"),
+           height = 10.5, width = 14.8, units = "cm")
 }
 
 omega_figure <- ggplot(filter(x, lambda == PARAMS$lambda, mu == PARAMS$mu, psi == PARAMS$psi, rho == PARAMS$rho, nu == PARAMS$nu), aes(x = omega, y = llhd)) +
     geom_line() +
     geom_vline(xintercept = PARAMS$omega, linetype = truth_linetype) +
     labs(x = "Occurrence rate",
-         y = "Log-Likelihood",
-         title = "Log-likelihood profile",
-         subtitle = "Known observation model parameters") +
+         y = "Log-Likelihood") +
+    y_scaling +
     fig_theme
 
 if (save_figures) {
-    ggsave("out/llhd-profile-omega.pdf", plot = omega_figure, height = 10.5, width = 14.8, units = "cm")
+    ggsave("out/llhd-profile-omega.pdf",
+           plot = omega_figure + labs(title = "Log-likelihood profile", subtitle = "Known observation model parameters"),
+           height = 10.5, width = 14.8, units = "cm")
 }
 
 nu_figure <- ggplot(filter(x, lambda == PARAMS$lambda, mu == PARAMS$mu, psi == PARAMS$psi, rho == PARAMS$rho, omega == PARAMS$omega), aes(x = nu, y = llhd)) +
     geom_line() +
     geom_vline(xintercept = PARAMS$nu, linetype = truth_linetype) +
-    labs(x = "Disaster extinction probability",
-         y = "Log-Likelihood",
-         title = "Log-likelihood profile",
-         subtitle = "Known observation model parameters") +
+    labs(x = "Disaster extinction\nprobability",
+         y = "Log-Likelihood") +
+    y_scaling +
     fig_theme
 
 if (save_figures) {
-    ggsave("out/llhd-profile-nu.pdf", plot = nu_figure, height = 10.5, width = 14.8, units = "cm")
+    ggsave("out/llhd-profile-nu.pdf",
+           plot = nu_figure + labs(title = "Log-likelihood profile", subtitle = "Known observation model parameters"),
+           height = 10.5, width = 14.8, units = "cm")
 }
 
 
-simulationEvents <- readLines(config$outputEventsFile)
+## Suppress warnings because otherwise there is a warning about a missing new
+## line character that does not affect things.
+simulationEvents <- suppressWarnings(readLines(config$outputEventsFile))
 et <- as.list(table(str_extract(simulationEvents, "^[a-zA-Z]+")))
 
 mask <- grepl(pattern = "CatastropheEvent", x = simulationEvents)
@@ -139,22 +150,41 @@ rm(mask)
 
 num_unobserved_lineages <- 1 + et$InfectionEvent - et$OccurrenceEvent - et$RemovalEvent - et$SamplingEvent - sum(num_catastropheed) - sum(num_disastered)
 
-unique_thresh <- 0.000001
+## Because we do not include the true parameters in the LLHD evaluations, we
+## need to get the closest match to them among those that we have.
+unique_thresh <- 0.0006
 curr_nb <- as.list(filter(x, abs(lambda - PARAMS$lambda) < unique_thresh, mu == PARAMS$mu, psi == PARAMS$psi, abs(rho - PARAMS$rho) < unique_thresh, omega == PARAMS$omega, abs(nu - PARAMS$nu) < unique_thresh) %>% select(starts_with("neg_binom")))
 curr_nb <- list(neg_binom_r = mean(curr_nb$neg_binom_r), neg_binom_p = mean(curr_nb$neg_binom_p))
 
-prev_mesh <- 1:100 * 10
+prev_mesh <- 130:350
 plot_df <- data.frame(prevalence = prev_mesh, log_prob = dnbinom(prev_mesh, size = curr_nb$neg_binom_r, prob = curr_nb$neg_binom_p, log = TRUE))
 
 prev_figure <- ggplot(plot_df, aes(x = prevalence, y = log_prob)) +
     geom_line() +
     geom_vline(xintercept = num_unobserved_lineages, linetype = truth_linetype) +
     labs(x = "Unobserved Lineages",
-         y = "Log-Likelihood",
-         title = "Log-likelihood profile",
-         subtitle = "Distribution of remaining lineages at present assuming known parameters") +
+         y = "Log-Likelihood") +
     fig_theme
 
 if (save_figures) {
     ggsave("out/llhd-profile-prevalence.pdf", plot = prev_figure, height = 10.5, width = 14.8, units = "cm")
+}
+
+
+## We want a plot combining the profiles for all of the parameters so we need to
+## combine them now.
+
+combined_llhd_profiles <- grid.arrange(lambda_figure,
+                                       mu_figure,
+                                       psi_figure,
+                                       rho_figure,
+                                       omega_figure,
+                                       nu_figure,
+                                       ncol=3,
+                                       left = "Log-Likelihood")
+
+if (save_figures) {
+    ggsave("out/parameter-llhd-profiles.pdf",
+           plot = combined_llhd_profiles,
+           height = 10.5, width = 14.8, units = "cm")
 }